+++ .gitignore
@@ -2,6 +2,7 @@
 
 # dependencies
 /node_modules
+/build
 /.pnp
 .pnp.js
