+++ src/components/ExternalImport/ORCID/OrcidList.jsx
@@ -1,20 +1,18 @@
 import React, { useContext, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { getRor } from "../../services/RorApi";
-import CustomError from "../Shared/CustomError";
-import CustomSpinner from "../Shared/CustomSpinner";
-import Pagination from "./Pagination";
-import Select from "react-select";
-import { GlobalContext } from "../context/Global";
+import { getOrcide } from "../../../services/RorApi";
+import { GlobalContext } from "../../context/Global";
+import CustomError from "../../Shared/CustomError";
+import CustomSpinner from "../../Shared/CustomSpinner";
+import Pagination from "../Pagination";
 
-function RorList() {
+function OrcidList() {
   const { t } = useTranslation();
   const [data, setData] = useState(null);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState(null);
   const [currentData, setcurrentData] = useState([]);
   const [allInitialData, setallInitialData] = useState([]);
-  const [contries, setContries] = useState([]);
   const [selectedKey, setSelectedKey] = useState(null);
   const { temp, setTemp } = useContext(GlobalContext);
 
@@ -31,23 +29,10 @@ component is initially rendered. */
    */
   const getData = () => {
     setLoading(true);
-    getRor()
+    getOrcide()
       .then((res) => {
         setData(res.data);
         setallInitialData(res.data);
-        const options = res.data.map((option) => ({
-          value: option.country.code,
-          label: option.country.name,
-          object: option,
-        }));
-        // get distinct array of objects
-        let distinctArr = Object.values(
-          options.reduce((acc, cur) => {
-            if (!acc[cur.value]) acc[cur.value] = cur;
-            return acc;
-          }, {})
-        );
-        setContries(distinctArr);
       })
       .catch((error) => setError(error))
       .finally(() => setLoading(false));
@@ -65,19 +50,10 @@ component is initially rendered. */
    * The function `setSelectedValue` updates the selected key and sets a temporary object with affiliation information.
    */
   const setSelectedValue = (el) => {
-    setSelectedKey(selectedKey === el.ror ? null : el.ror);
-    const obj = { affiliationId: el.ror, affiliationName: el.name[Object.keys(el.name)[0]] };
+    console.log(el);
+    setSelectedKey(selectedKey === el.orcid ? null : el.orcid);
+    const obj = { firstName: el.givenNames, lastName: el?.familyNames, personId: el.orcid, nameType: "Personne", idType: "ORCID iD" };
     setTemp({ ...temp, ...obj });
-  };
-
-  /**
-   * The handleChangeCounty function filters an array of data based on the selected country code and updates the data state.
-   */
-  const handleChangeCounty = (e) => {
-    const filterPays = allInitialData.filter((el) => {
-      return el.country.code == e.value;
-    });
-    setData(filterPays);
   };
 
   /**
@@ -86,7 +62,11 @@ component is initially rendered. */
   const handleChangeText = (e) => {
     const text = e.target.value;
     const filterText = allInitialData.filter((el) => {
-      return el.name[Object.keys(el.name)[0]].toLowerCase().includes(text.toLowerCase()) || el.ror.toLowerCase().includes(text.toLowerCase());
+      return (
+        el.familyNames.toLowerCase().includes(text.toLowerCase()) ||
+        el.givenNames.toLowerCase().includes(text.toLowerCase()) ||
+        el.orcid.toLowerCase().includes(text.toLowerCase())
+      );
     });
     setData(filterText);
   };
@@ -97,28 +77,6 @@ component is initially rendered. */
       {!loading && error && <CustomError></CustomError>}
       {!loading && !error && data && (
         <>
-          <div className="row" style={{ margin: "10px" }}>
-            <div className="">
-              <div className="row">
-                <div>
-                  <Select
-                    menuPortalTarget={document.body}
-                    styles={{
-                      menuPortal: (base) => ({ ...base, zIndex: 9999 }),
-                      singleValue: (base) => ({ ...base, color: "var(--primary)" }),
-                      control: (base) => ({ ...base, borderRadius: "8px", borderWidth: "1px", borderColor: "var(--primary)", height: "43px" }),
-                    }}
-                    onChange={handleChangeCounty}
-                    defaultValue={{
-                      label: t("Select a country"),
-                      value: t("Select a country"),
-                    }}
-                    options={contries}
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
           <div className="row" style={{ margin: "10px" }}>
             <div>
               <div className="row">
@@ -150,43 +108,34 @@ component is initially rendered. */
             <table className="table table-bordered table-hover">
               <thead className="thead-dark">
                 <tr>
-                  <th scope="col">{t("Organization name")}</th>
-                  <th scope="col">{t("Acronym")}</th>
-                  <th scope="col">{t("Country")}</th>
-                  <th scope="col">{t("Location")}</th>
-                  <th scope="col">ROR</th>
+                  <th scope="col">{t("ID")}</th>
+                  <th scope="col">{t("Last / First name")}</th>
+                  <th scope="col">{t("Establishment")}</th>
                   <th scope="col"></th>
                 </tr>
               </thead>
               <tbody>
                 {currentData.map((el, idx) => (
                   <tr key={idx}>
+                    <td scope="row">{el.orcid}</td>
+                    <td>{`${el.familyNames} ${el.givenNames} `}</td>
                     <td scope="row">
-                      <a href={el.links[0]} target="_blank" rel="noopener noreferrer">
-                        {el.name[Object.keys(el.name)[0]]}
-                      </a>
+                      {el?.institutionName.map((e, idx) => (
+                        <React.Fragment key={idx}>{e}</React.Fragment>
+                      ))}
                     </td>
-                    <td>{el.acronyms}</td>
-                    <td>{el.country.code}</td>
-                    <td scope="row">
-                      {Object.values(el.addresses[0])
-                        .filter((value) => value)
-                        .join(", ")}
-                    </td>
-                    <td>{el.ror}</td>
                     <td>
-                      <input className="text-center" type="checkbox" checked={selectedKey === el.ror} onChange={() => setSelectedValue(el)} />
+                      <input className="text-center" type="checkbox" checked={selectedKey === el.orcid} onChange={() => setSelectedValue(el)} />
                     </td>
                   </tr>
                 ))}
               </tbody>
             </table>
           </div>
-
           <div className="row text-center">
             <div className="mx-auto"></div>
             <div className="mx-auto">
-              <Pagination items={data} onChangePage={onChangePage} pageSize={8} />
+              <Pagination items={data} onChangePage={onChangePage} pageSize={5} />
             </div>
           </div>
         </>
@@ -195,4 +144,4 @@ component is initially rendered. */
   );
 }
 
-export default RorList;
+export default OrcidList;
