+++ src/components/Forms/SelectSingleList.jsx
@@ -5,8 +5,6 @@ import { getDefaultLabel } from "../../utils/GeneratorUtils";
 import { GlobalContext } from "../context/Global";
 import styles from "../assets/css/form.module.css";
 import { useTranslation } from "react-i18next";
-import RorList from "../ROR/RorList";
-import OrcidList from "../ORCID/OrcidList";
 
 /* This is a functional component in JavaScript React that renders a select list with options fetched from a registry. It takes in several props such as
 label, name, changeValue, tooltip, registry, and schemaId. It uses the useState and useEffect hooks to manage the state of the options and to fetch
@@ -15,17 +13,24 @@ the list, and it updates the value of the input field accordingly. Finally, it r
 function SelectSingleList({ label, name, changeValue, tooltip, registry, schemaId, readonly, registries }) {
   const { t, i18n } = useTranslation();
   const [lng] = useState(i18n.language.split("-")[0]);
-  const [options, setoptions] = useState(null);
   const { form, temp } = useContext(GlobalContext);
   const [registryName, setRegistryName] = useState(registry);
   const [selectMonted, setSelectMonted] = useState(true);
-  const [showRor, setShowRor] = useState(false);
-  const [showOrcid, setShowOrcid] = useState(false);
-  const [renderKey, setRenderKey] = useState(0);
+  const [options, setoptions] = useState([]); // Initialized to an empty array
+  const [currentValue, setCurrentValue] = useState({ label: "", value: "" }); // Initialized to an empty object with label and value
+
+  /* The `useEffect` hook is used to perform side effects in functional components. In this case, the effect is triggered when the `temp` variable
+ changes. */
+  useEffect(() => {
+    const newDefaultValue = {
+      label: getDefaultLabel(temp, form?.[schemaId], name) || "", // Use default value if undefined
+      value: getDefaultLabel(temp, form?.[schemaId], name) || "", // Use default value if undefined
+    };
+    setCurrentValue(newDefaultValue);
+  }, [temp]);
 
   /* A hook that is called when the component is mounted. It is used to set the options of the select list. */
   useEffect(() => {
-    setRenderKey((prevKey) => prevKey + 1);
     let isMounted = true;
     const createOptions = (data) => {
       return data.map((option) => ({
@@ -52,26 +57,17 @@ function SelectSingleList({ label, name, changeValue, tooltip, registry, schemaI
     return () => {
       isMounted = false;
     };
-  }, [registryName, lng]);
+  }, [registryName, lng, temp]);
 
   /**
    * It takes the value of the input field and adds it to the list array.
    * @param e - the event object
    */
   const handleChangeList = (e) => {
-    const isRorID = e.value === "ROR ID";
-    const isOrcidID = e.value === "ORCID iD";
-
-    setShowRor(isRorID);
-    setShowOrcid(isOrcidID);
-
-    if (!isRorID && !isOrcidID) {
-      setShowRor(false);
-      setShowOrcid(false);
-    }
-
     const targetValue = name === "funder" ? e.object : e.value;
     changeValue({ target: { name, value: targetValue } });
+    console.log(e);
+    setCurrentValue({ label: e.label, value: e.value });
   };
 
   /**
@@ -129,25 +125,20 @@ function SelectSingleList({ label, name, changeValue, tooltip, registry, schemaI
               <div className="row">
                 <div className={`col-md-12 ${styles.select_wrapper}`}>
                   {selectMonted && options && (
-                    <>
-                      <Select
-                        menuPortalTarget={document.body}
-                        styles={{
-                          menuPortal: (base) => ({ ...base, zIndex: 9999, color: "grey" }),
-                          singleValue: (base) => ({ ...base, color: "var(--primary)" }),
-                          control: (base) => ({ ...base, borderRadius: "8px", borderWidth: "1px", borderColor: "var(--primary)" }),
-                        }}
-                        onChange={handleChangeList}
-                        options={options}
-                        name={name}
-                        style={{ color: "red" }}
-                        defaultValue={{
-                          label: getDefaultLabel(temp, form?.[schemaId], name),
-                          value: getDefaultLabel(temp, form?.[schemaId], name),
-                        }}
-                        isDisabled={readonly}
-                      />
-                    </>
+                    <Select
+                      menuPortalTarget={document.body}
+                      styles={{
+                        menuPortal: (base) => ({ ...base, zIndex: 9999, color: "grey" }),
+                        singleValue: (base) => ({ ...base, color: "var(--primary)" }),
+                        control: (base) => ({ ...base, borderRadius: "8px", borderWidth: "1px", borderColor: "var(--primary)" }),
+                      }}
+                      onChange={handleChangeList}
+                      options={options}
+                      name={name}
+                      style={{ color: "red" }}
+                      value={currentValue}
+                      isDisabled={readonly}
+                    />
                   )}
                 </div>
               </div>
@@ -156,28 +147,6 @@ function SelectSingleList({ label, name, changeValue, tooltip, registry, schemaI
         </div>
         {/* *************Select ref************* */}
       </div>
-      <React.Fragment key={renderKey + 1}>
-        {showOrcid && (
-          <>
-            <OrcidList></OrcidList>
-            <div style={{ display: "flex", justifyContent: "center" }}>
-              {/* <Button variant="secondary" onClick={handleClose}>
-              {t("Close")}
-            </Button> */}
-            </div>
-          </>
-        )}
-        {showRor && (
-          <>
-            <RorList></RorList>
-            <div style={{ display: "flex", justifyContent: "center" }}>
-              {/* <Button variant="secondary" onClick={handleClose}>
-              {t("Close")}
-            </Button> */}
-            </div>
-          </>
-        )}
-      </React.Fragment>
     </>
   );
 }
